+++ src/reference-implementations/contracts/reference-implementation-methodology-profile.js
@@ -1,0 +1,33 @@
+export class ReferenceImplementationMethodologyProfile {
+  constructor({
+    methodologyName,
+    methodPrinciples = [], scenarioSelectionCriteria = [], scenarioDefinitionFields = [],
+    architectureTraceabilityMappings = [], conformanceProfileElements = [], implementationLifecycleSteps = [],
+    evidencePackageItems = [], variationPointCategories = [], divergenceClassifications = [], reviewGateAreas = [],
+    maintenanceMonitoringAreas = [], retirementElements = [], architecturalRules = [],
+    traceabilityIsBidirectionalBetweenArchitectureAndExamples = true,
+    substitutionRetainsContractsInvariantsTestsQualityOutcomes = true,
+    newScenarioMovesFromProposalToReleaseWithRepeatableGatesEvidenceOwnership = true,
+    conformanceIsProfileBasedAndEvidenceBacked = true, traceabilityIsMandatoryForMaterialArtifacts = true,
+    variationIsExplicitAndTestPreserving = true, maintenanceAndRetirementArePartOfReferenceScope = true,
+    architectureChangesFollowGovernanceNeverExampleDrift = true,
+    scenariosDuplicateWithoutDemonstratingMeaningfulVariant = false, unclassifiedDivergencePresentedAsConformant = false,
+    staleReferenceSilentlyPresentedAsCurrent = false
+  }) {
+    this.methodologyName = methodologyName;
+    for (const [key, value] of Object.entries({
+      methodPrinciples, scenarioSelectionCriteria, scenarioDefinitionFields, architectureTraceabilityMappings,
+      conformanceProfileElements, implementationLifecycleSteps, evidencePackageItems, variationPointCategories,
+      divergenceClassifications, reviewGateAreas, maintenanceMonitoringAreas, retirementElements, architecturalRules
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      traceabilityIsBidirectionalBetweenArchitectureAndExamples, substitutionRetainsContractsInvariantsTestsQualityOutcomes,
+      newScenarioMovesFromProposalToReleaseWithRepeatableGatesEvidenceOwnership, conformanceIsProfileBasedAndEvidenceBacked,
+      traceabilityIsMandatoryForMaterialArtifacts, variationIsExplicitAndTestPreserving,
+      maintenanceAndRetirementArePartOfReferenceScope, architectureChangesFollowGovernanceNeverExampleDrift,
+      scenariosDuplicateWithoutDemonstratingMeaningfulVariant, unclassifiedDivergencePresentedAsConformant,
+      staleReferenceSilentlyPresentedAsCurrent
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}