+++ src/reference-implementations/service-registration.js
@@ -1,8 +1,10 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { ReferenceImplementationOverviewDescriptor } from './overview/reference-implementation-overview-descriptor.js';
+import { ReferenceImplementationMethodologyDescriptor } from './methodology/reference-implementation-methodology-descriptor.js';
 
 export function addReferenceImplementations(services) {
   if (!(services instanceof ServiceCollection)) throw new Error('addReferenceImplementations expects an instance of ServiceCollection.');
   services.registerSingleton('ReferenceImplementationOverviewDescriptor', () => new ReferenceImplementationOverviewDescriptor());
+  services.registerSingleton('ReferenceImplementationMethodologyDescriptor', () => new ReferenceImplementationMethodologyDescriptor());
   return services;
 }